Change default workremaining to 0 instead of 1 for improving downscaling
